Fix a compile warning on 64bit OS X
Backport of 68475fc5c5a806ebbb5657de1667dab2c3e09b7c which accidentally
only made it into master. Fixes bug 4547. Bug isn't in any released
version.

+++ src/or/dirserv.c
@@ -1909,7 +1909,7 @@ dirserv_compute_performance_thresholds(routerlist_t *rl)
    * of Fast nodes. */
   if (options->AuthDirFastGuarantee &&
       fast_bandwidth > options->AuthDirFastGuarantee)
-    fast_bandwidth = options->AuthDirFastGuarantee;
+    fast_bandwidth = (uint32_t)options->AuthDirFastGuarantee;
 
   /* Now that we have a time-known that 7/8 routers are known longer than,
    * fill wfus with the wfu of every such "familiar" router. */
